Pre-parse public keys into an optimised representation
