+++ src/core/crypto/CMakeLists.txt
@@ -12,8 +12,7 @@ if(SOURCEMETA_CORE_CRYPTO_USE_SYSTEM_OPENSSL)
   target_sources(sourcemeta_core_crypto PRIVATE
     crypto_sha1_openssl.cc crypto_sha256_openssl.cc crypto_sha384_openssl.cc
     crypto_sha512_openssl.cc crypto_random_openssl.cc
-    crypto_verify_rsa_openssl.cc crypto_verify_ecdsa_openssl.cc
-    crypto_verify_eddsa_openssl.cc)
+    crypto_publickey_openssl.cc)
   target_link_libraries(sourcemeta_core_crypto PRIVATE OpenSSL::Crypto)
 elseif(APPLE)
   enable_language(OBJCXX)
@@ -99,8 +98,8 @@ elseif(APPLE)
   target_sources(sourcemeta_core_crypto PRIVATE
     crypto_sha1_apple.cc crypto_sha256_apple.cc crypto_sha384_apple.cc
     crypto_sha512_apple.cc crypto_random_apple.cc
-    crypto_verify_rsa_apple.cc crypto_verify_ecdsa_apple.cc
-    crypto_verify_eddsa_apple.cc crypto_eddsa_cryptokit.mm "${CRYPTOKIT_OBJECT}"
+    crypto_publickey_apple.cc
+    crypto_eddsa_cryptokit.mm "${CRYPTOKIT_OBJECT}"
     crypto_eddsa.h crypto_eddsa_apple.h crypto_bignum.h crypto_shake256.h)
 
   # The generated Objective-C interface header lives in the build tree
@@ -124,18 +123,15 @@ elseif(WIN32)
   target_sources(sourcemeta_core_crypto PRIVATE
     crypto_sha1_windows.cc crypto_sha256_windows.cc crypto_sha384_windows.cc
     crypto_sha512_windows.cc crypto_random_windows.cc
-    crypto_verify_rsa_windows.cc crypto_verify_ecdsa_windows.cc
-    crypto_verify_eddsa_windows.cc crypto_eddsa.h crypto_bignum.h
-    crypto_shake256.h)
+    crypto_publickey_windows.cc crypto_eddsa.h
+    crypto_bignum.h crypto_shake256.h)
   target_link_libraries(sourcemeta_core_crypto PRIVATE bcrypt)
 else()
   message(WARNING "Building the reference cryptography backend, instead of the "
     "OpenSSL recommended production one")
   target_sources(sourcemeta_core_crypto PRIVATE
     crypto_sha1_other.cc crypto_sha256_other.cc crypto_sha384_other.cc
-    crypto_sha512_other.cc crypto_random_other.cc
-    crypto_verify_rsa_other.cc crypto_verify_ecdsa_other.cc
-    crypto_verify_eddsa_other.cc
+    crypto_sha512_other.cc crypto_random_other.cc crypto_publickey_other.cc
     crypto_bignum.h crypto_ecc.h crypto_eddsa.h crypto_shake256.h)
 endif()
