--- src/core/crypto/crypto_verify_eddsa_windows.cc
@@ -1,27 +1,0 @@
-#include <sourcemeta/core/crypto_verify.h>
-
-#include "crypto_eddsa.h"
-
-#include <string_view> // std::string_view
-#include <utility>     // std::unreachable
-
-// Windows CNG exposes no EdDSA primitive through its C API, so this backend
-// verifies with the reference implementation rather than the platform, as the
-// other primitives do
-
-namespace sourcemeta::core {
-
-auto eddsa_verify(const EdwardsCurve curve, const std::string_view public_key,
-                  const std::string_view message,
-                  const std::string_view signature) -> bool {
-  switch (curve) {
-    case EdwardsCurve::Ed25519:
-      return edwards25519_verify(public_key, message, signature);
-    case EdwardsCurve::Ed448:
-      return edwards448_verify(public_key, message, signature);
-  }
-
-  std::unreachable();
-}
-
-} // namespace sourcemeta::core
