+++ src/core/jose/CMakeLists.txt
@@ -7,7 +7,7 @@ sourcemeta_library(NAMESPACE sourcemeta PROJECT core NAME jose
 target_link_libraries(sourcemeta_core_jose
   PUBLIC sourcemeta::core::json)
 target_link_libraries(sourcemeta_core_jose
-  PRIVATE sourcemeta::core::crypto)
+  PUBLIC sourcemeta::core::crypto)
 target_link_libraries(sourcemeta_core_jose
   PRIVATE sourcemeta::core::text)
 target_link_libraries(sourcemeta_core_jose
