+++ src/core/jose/jose_jws_verify_signature.cc
@@ -52,56 +52,47 @@ auto jws_verify_signature(const std::optional<JWSAlgorithm> algorithm,
     return false;
   }
 
+  // The key material is parsed into a reusable platform key when the key is
+  // constructed, so an absent one is material that never formed a valid key
+  const auto *public_key{key.public_key()};
+  if (public_key == nullptr) {
+    return false;
+  }
+
   switch (algorithm.value()) {
     case JWSAlgorithm::RS256:
     case JWSAlgorithm::RS384:
     case JWSAlgorithm::RS512:
       return key.type() == JWK::Type::RSA &&
-             rsassa_pkcs1_v15_verify(hash_for(algorithm.value()), key.modulus(),
-                                     key.exponent(), signing_input, signature);
+             rsassa_pkcs1_v15_verify(*public_key, hash_for(algorithm.value()),
+                                     signing_input, signature);
     case JWSAlgorithm::PS256:
     case JWSAlgorithm::PS384:
     case JWSAlgorithm::PS512:
       return key.type() == JWK::Type::RSA &&
-             rsassa_pss_verify(hash_for(algorithm.value()), key.modulus(),
-                               key.exponent(), signing_input, signature);
+             rsassa_pss_verify(*public_key, hash_for(algorithm.value()),
+                               signing_input, signature);
     // Each ECDSA algorithm is pinned to exactly one curve (RFC 7518 Section
     // 3.4), so the key's curve is checked independently of any algorithm it
     // declares
     case JWSAlgorithm::ES256:
       return key.type() == JWK::Type::EllipticCurve && key.curve() == "P-256" &&
-             ecdsa_verify(EllipticCurve::P256, SignatureHashFunction::SHA256,
-                          key.coordinate_x(), key.coordinate_y(), signing_input,
-                          signature);
+             ecdsa_verify(*public_key, SignatureHashFunction::SHA256,
+                          signing_input, signature);
     case JWSAlgorithm::ES384:
       return key.type() == JWK::Type::EllipticCurve && key.curve() == "P-384" &&
-             ecdsa_verify(EllipticCurve::P384, SignatureHashFunction::SHA384,
-                          key.coordinate_x(), key.coordinate_y(), signing_input,
-                          signature);
+             ecdsa_verify(*public_key, SignatureHashFunction::SHA384,
+                          signing_input, signature);
     case JWSAlgorithm::ES512:
       return key.type() == JWK::Type::EllipticCurve && key.curve() == "P-521" &&
-             ecdsa_verify(EllipticCurve::P521, SignatureHashFunction::SHA512,
-                          key.coordinate_x(), key.coordinate_y(), signing_input,
-                          signature);
+             ecdsa_verify(*public_key, SignatureHashFunction::SHA512,
+                          signing_input, signature);
     // The Edwards-curve algorithm names one of two curves through the key
-    // rather than the algorithm (RFC 8037 Section 3.1), and the public key is a
-    // single encoded point exposed as the first coordinate
+    // rather than the algorithm (RFC 8037 Section 3.1), and the key fixes the
+    // curve when it is parsed
     case JWSAlgorithm::EdDSA:
-      if (key.type() != JWK::Type::OctetKeyPair) {
-        return false;
-      }
-
-      if (key.curve() == "Ed25519") {
-        return eddsa_verify(EdwardsCurve::Ed25519, key.coordinate_x(),
-                            signing_input, signature);
-      }
-
-      if (key.curve() == "Ed448") {
-        return eddsa_verify(EdwardsCurve::Ed448, key.coordinate_x(),
-                            signing_input, signature);
-      }
-
-      return false;
+      return key.type() == JWK::Type::OctetKeyPair &&
+             eddsa_verify(*public_key, signing_input, signature);
   }
 
   std::unreachable();
